Move App route checks into dedicated test file
