+++ ui/src/App.tsx
@@ -334,53 +334,3 @@ export function App() {
     </>
   );
 }
-
-type RouteLike = {
-  props?: {
-    path?: string;
-    children?: unknown;
-  };
-};
-
-function collectRoutePaths(node: unknown, paths: string[] = []): string[] {
-  if (!node || typeof node !== "object") return paths;
-
-  const route = node as RouteLike;
-  if (route.props?.path) paths.push(route.props.path);
-
-  const children = route.props?.children;
-  if (Array.isArray(children)) {
-    for (const child of children) {
-      collectRoutePaths(child, paths);
-    }
-  } else if (children) {
-    collectRoutePaths(children, paths);
-  }
-
-  return paths;
-}
-
-const testApi = (import.meta as ImportMeta & { vitest?: unknown }).vitest as
-  | {
-      describe: (name: string, fn: () => void) => void;
-      it: (name: string, fn: () => void) => void;
-      expect: (value: unknown) => { toContain: (item: string) => void };
-    }
-  | undefined;
-
-if (testApi) {
-  const { describe, it, expect } = testApi;
-
-  describe("App routes", () => {
-    it("includes project schedules under company-prefixed board routes", () => {
-      const paths = collectRoutePaths(boardRoutes());
-      expect(paths).toContain("projects/:projectId/schedules");
-    });
-
-    it("includes unprefixed redirects for project schedules and budget", () => {
-      const paths = collectRoutePaths(App());
-      expect(paths).toContain("projects/:projectId/schedules");
-      expect(paths).toContain("projects/:projectId/budget");
-    });
-  });
-}
